refactor: move class CustomIR into the cpp file to hide it
