+++ LR2/LR2_customir.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "LR2_customir_api.h"
-
 #include <array>
 #include <filesystem>
 #include <future>
@@ -9,28 +7,9 @@
 #include <string>
 #include <type_traits>
 
-#include <wtypes.h>
-
 struct game;
 struct sqlite3;
-
-class CustomIR {
-public:
-	CustomIR() = delete;
-	CustomIR(const std::filesystem::path& directory);
-	bool Initialize();
-	bool Login();
-	SendScoreStatus SendScore(const IRScoreV1& score);
-
-	[[nodiscard]] const std::string& Name() const { return mName; };
-private:
-	struct ModuleDeleter {
-		void operator()(std::remove_pointer_t<HMODULE>* handle);
-	};
-	std::unique_ptr<std::remove_pointer_t<HMODULE>, ModuleDeleter> mDllHandle;
-	std::string mName;
-	MethodTable mMethods;
-};
+class CustomIR;
 
 class CUSTOMIR_MANAGER {
 public:
